feat: remove fe
