+++ steps/common_io.py
@@ -2,11 +2,12 @@
 A set of common IO functions to be used across all steps.
 """
 
-import os
 import logging
 import shutil
-import conf
+from pathlib import Path
 
+import boto3
+import conf
 import polars as pl
 
 
@@ -19,15 +20,16 @@ def io_create_root_data_folder() -> bool:
         bool: True if all folders are created successfully, False otherwise.
     """
     try:
-        if not os.path.exists(conf.root_data_folder):
-            os.makedirs(conf.root_data_folder, exist_ok=True)
+        if not Path(conf.root_data_folder).exists():
+            Path(conf.root_data_folder).mkdir(exist_ok=True)
             logging.info(f"Created root data folder: {conf.root_data_folder}")
         else:
             logging.info(f"Root data folder already exists: {conf.root_data_folder}")
 
         for folder in conf.OutputPathType:
-            folder_path = os.path.join(conf.root_data_folder, folder.value)
-            os.makedirs(folder_path, exist_ok=True)
+            folder_path = Path(conf.root_data_folder) / Path(folder.value)
+            # folder_path = os.path.join(conf.root_data_folder, folder.value)
+            Path(folder_path).mkdir(exist_ok=True)
             logging.info(f"Created {folder.value} folder")
 
         return True
@@ -44,7 +46,7 @@ def io_remove_root_data_folder() -> bool:
         bool: True if the folder is deleted successfully, False otherwise.
     """
     try:
-        if os.path.exists(conf.root_data_folder):
+        if Path(conf.root_data_folder).exists():
             shutil.rmtree(conf.root_data_folder)
             logging.info(f"Files from '{conf.root_data_folder}' removed")
             return True
@@ -81,3 +83,29 @@ def io_write_from_local_to_s3(df: pl.DataFrame, local_path: str, s3_key: str, de
     else:
         logging.info("Cleaning up local files")
         io_remove_root_data_folder()
+
+
+def upload_reference_tables_to_s3(s3_client: boto3.client, parquet_files: list[str]) -> None:
+    """
+    Upload Parquet reference tables to the configured S3 bucket.
+
+    Args:
+        s3_client (boto3.client): Boto3 S3 client instance.
+        parquet_files (list[str]): List of local Parquet file paths to upload.
+
+    Returns:
+        None
+    """
+    logging.info(f"Uploading {len(parquet_files)} reference tables to S3")
+
+    for file in parquet_files:
+        paruqet_file_name = file.split("/")[-1]
+        reference_s3_key = conf.create_output_file_path(conf.OutputPathType.REFERENCE, conf.FileLocation.S3, file_name=paruqet_file_name)
+
+        s3_client.upload_file(
+            Filename=file,
+            Bucket=conf.bucket_name,
+            Key=reference_s3_key,
+        )
+
+        logging.info(f"Uploaded {paruqet_file_name} to s3://{conf.bucket_name}/{reference_s3_key}")
